# swagger.yaml
openapi: 3.0.3
...
info:
  title: Laundry Service API
  description: |
    ## 🧺 Tổng quan
    API cho nền tảng dịch vụ Giặt ủi trực tuyến. Hệ thống hỗ trợ:
    - **Khách hàng (Customer)**: Đặt đơn giặt, theo dõi trạng thái, thanh toán
    - **Nhân viên (Staff)**: Xử lý đơn hàng, cập nhật trạng thái, upload ảnh
    - **Quản trị viên (Admin)**: Toàn quyền quản lý hệ thống

    ## 🔐 Xác thực
    Hệ thống hỗ trợ **2 luồng đăng nhập**:

    ### Luồng 1: Email + Password (dành cho Admin & Staff)
    ```
    POST /auth/register → POST /auth/login → nhận { accessToken }
    ```

    ### Luồng 2: SĐT + OTP (dành cho Customer — không cần password)
    ```
    POST /auth/send-otp → POST /auth/verify-otp → nhận { accessToken }
    ```

    ### Cách truyền Token lên Header
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    ```

    ### Refresh Token
    - Được lưu trong **HttpOnly Cookie** (`refreshToken`), browser tự gửi kèm
    - Khi access token hết hạn, gọi `POST /auth/refresh-token` để cấp lại

    ## 📦 Luồng trạng thái đơn hàng
    ```
    received → washing → drying → delivering → completed
        ↓          ↓        ↓           ↓
    cancelled  cancelled cancelled  cancelled
    ```
    *(Customer chỉ hủy được khi status = received. Admin hủy được mọi lúc trừ completed)*

    ## 📝 Cấu trúc Response chuẩn
    Mọi response đều theo format:
    ```json
    {
      "success": true,
      "message": "Mô tả kết quả",
      "data": { ... },
...
    }
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token. Lấy từ response của `/auth/login` hoặc `/auth/verify-otp`.
        Truyền vào header: `Authorization: Bearer <token>`

  # ═══════════════════════════════════════════════════════════════════════════
  #  SCHEMAS (Data Models / Types)
  #  FE có thể dùng phần này để tạo TypeScript Interface/Type
  # ═══════════════════════════════════════════════════════════════════════════
  schemas:
...
    Service:
      type: object
      description: Gói dịch vụ giặt ủi
      properties:
        _id:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d2"
        name:
          type: string
          example: "Giặt + Sấy thường"
          description: Tên dịch vụ (unique)
        description:
          type: string
          example: "Giặt máy kết hợp sấy khô, phù hợp quần áo hàng ngày"
          nullable: true
        priceType:
          type: string
          enum: [per_kg, per_item]
          example: "per_kg"
          description: |
            Loại tính giá:
            - `per_kg`: Tính theo kg (vd: 25.000đ/kg)
            - `per_item`: Tính theo món (vd: 80.000đ/món áo vest)
        price:
          type: number
          example: 25000
          description: Đơn giá (VNĐ). Tổng tiền = price × quantity
        estimatedHours:
          type: number
          example: 24
          description: Thời gian xử lý ước tính (đơn vị giờ)
        isActive:
          type: boolean
          example: true
          description: Dịch vụ có đang được cung cấp không
        createdAt:
          type: string
          format: date-time
          example: "2024-01-10T08:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-06-01T09:00:00.000Z"

    # ─── STATUS HISTORY (Lịch sử trạng thái đơn hàng) ─────────────────────────
...
    Payment:
      type: object
      description: Giao dịch thanh toán
      properties:
        _id:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d6"
        order:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d3"
          description: OrderId
        amount:
          type: number
          example: 87500
          description: Số tiền thanh toán (VNĐ)
        method:
          type: string
          enum: [cash, vnpay, momo]
          example: "vnpay"
          description: Phương thức thanh toán
        status:
          type: string
          enum: [pending, paid, failed, refunded]
          example: "pending"
          description: |
            Trạng thái giao dịch:
            - `pending`: Chờ thanh toán
            - `paid`: Đã thanh toán thành công
            - `failed`: Thanh toán thất bại
            - `refunded`: Đã hoàn tiền
        transactionId:
          type: string
          example: "VNP20240620090000123456"
          description: Mã giao dịch từ cổng thanh toán
          nullable: true
        paymentUrl:
          type: string
          format: uri
          example: "https://sandbox.vnpayment.vn/paymentv2/vpcpay.html?vnp_Amount=8750000&..."
          description: URL redirect để khách thanh toán (chỉ có với vnpay/momo)
          nullable: true
        paidAt:
          type: string
          format: date-time
          example: "2024-06-20T09:05:00.000Z"
          description: Thời điểm thanh toán thành công
          nullable: true
        createdAt:
          type: string
          format: date-time
          example: "2024-06-20T09:00:00.000Z"

    # ─── DASHBOARD STATS ──────────────────────────────────────────────────────
...
    CreateServiceRequest:
      type: object
      required: [name, priceType, price]
      description: Body để tạo gói dịch vụ mới
      properties:
        name:
          type: string
          example: "Giặt khô cao cấp"
          description: Tên dịch vụ (phải là duy nhất trong hệ thống)
        description:
          type: string
          example: "Giặt khô chuyên nghiệp, phù hợp vest, áo khoác da, đầm dạ hội"
          nullable: true
        priceType:
          type: string
          enum: [per_kg, per_item]
          example: "per_item"
        price:
          type: number
          example: 150000
          minimum: 0
          description: Đơn giá (VNĐ)
        estimatedHours:
          type: number
          example: 48
          minimum: 1
          description: Thời gian xử lý ước tính (giờ, mặc định 24)
    UpdateServiceRequest:
      type: object
      description: Body để cập nhật gói dịch vụ (chỉ truyền các field cần thay đổi)
      properties:
        name:
          type: string
          example: "Giặt khô cao cấp (updated)"
        description:
          type: string
          nullable: true
        priceType:
          type: string
          enum: [per_kg, per_item]
        price:
          type: number
          minimum: 0
        estimatedHours:
          type: number
          minimum: 1
        isActive:
          type: boolean
          example: false
          description: Đặt false để tạm ngừng cung cấp dịch vụ
    CreatePaymentRequest:
      type: object
      required: [orderId, method]
      description: Body để tạo giao dịch thanh toán
      properties:
        orderId:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d3"
          description: ID đơn hàng cần thanh toán
        method:
          type: string
          enum: [cash, vnpay, momo]
          example: "vnpay"
          description: Phương thức thanh toán
...
    ServiceListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Lấy danh sách dịch vụ thành công"
        data:
          type: object
          properties:
            services:
              type: array
              items:
                $ref: '#/components/schemas/Service'
    ServiceDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Lấy chi tiết dịch vụ thành công"
        data:
          type: object
          properties:
            service:
              $ref: '#/components/schemas/Service'
...
    ErrorResponse:
      type: object
      description: Cấu trúc lỗi chung
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Mô tả lỗi cụ thể"
        data:
          nullable: true
          example: null
    ValidationErrorResponse:
      type: object
      description: Lỗi validation (400) — khi dữ liệu đầu vào không hợp lệ
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Dữ liệu không hợp lệ"
        errors:
          type: array
          items:
            type: string
          example: ["Tên dịch vụ không được để trống", "Giá tiền không được âm"]
        data:
          nullable: true
          example: null

  # ─── REUSABLE PARAMETERS ──────────────────────────────────────────────────
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: MongoDB ObjectId của bản ghi
      schema:
        type: string
        example: "64f1a2b3c4d5e6f7a8b9c0d3"
...
  responses:
    Unauthorized:
      description: Token không hợp lệ hoặc chưa đăng nhập
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Bạn chưa đăng nhập. Vui lòng đăng nhập để tiếp tục."
            data: null
    Forbidden:
      description: Không có quyền truy cập
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Bạn không có quyền thực hiện hành động này."
            data: null
    NotFound:
      description: Không tìm thấy bản ghi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Không tìm thấy bản ghi"
            data: null
    BadRequest:
      description: Dữ liệu đầu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    InternalServerError:
      description: Lỗi server nội bộ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Lỗi server nội bộ"
            data: null

# ─── GLOBAL SECURITY ────────────────────────────────────────────────────────
# Áp dụng BearerAuth cho toàn bộ API.
# Các endpoint PUBLIC (không cần token) sẽ ghi đè bằng security: []
security:
  - BearerAuth: []

# ═══════════════════════════════════════════════════════════════════════════════
#  TAGS — Nhóm các endpoint
# ═══════════════════════════════════════════════════════════════════════════════
...
paths:
...
  /services:
    get:
      tags: [Services]
      summary: Lấy danh sách tất cả gói dịch vụ giặt ủi
      description: |
        Trả về danh sách các gói dịch vụ đang hoạt động.
        **Endpoint public — không cần token.**

        Ví dụ các gói dịch vụ:
        - Giặt + Sấy thường (25.000đ/kg)
        - Giặt khô áo vest (150.000đ/món)
        - Combo 5kg (110.000đ/combo)
      operationId: getAllServices
      security: []
      responses:
        '200':
          description: Lấy danh sách thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceListResponse'
              example:
                success: true
                message: "Lấy danh sách dịch vụ thành công"
                data:
                  services:
                    - _id: "64f1a2b3c4d5e6f7a8b9c0d2"
                      name: "Giặt + Sấy thường"
                      description: "Giặt máy kết hợp sấy khô, phù hợp quần áo hàng ngày"
                      priceType: "per_kg"
                      price: 25000
                      estimatedHours: 24
                      isActive: true
                    - _id: "64f1a2b3c4d5e6f7a8b9c0e1"
                      name: "Giặt khô cao cấp"
                      description: "Phù hợp vest, áo khoác, đầm dạ hội"
                      priceType: "per_item"
                      price: 150000
                      estimatedHours: 48
                      isActive: true
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Admin — Services Management]
      summary: "[Admin] Tạo gói dịch vụ mới"
      description: |
        **[Yêu cầu role: admin]**

        Tạo một gói dịch vụ giặt ủi mới.
        Tên dịch vụ phải là duy nhất trong hệ thống.
      operationId: createService
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
            example:
              name: "Giặt + Ủi phẳng"
              description: "Giặt máy và ủi phẳng từng món, giao hàng trong 36h"
              priceType: "per_kg"
              price: 35000
              estimatedHours: 36
      responses:
        '201':
          description: Tạo dịch vụ thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /services/{id}:
    get:
      tags: [Services]
      summary: Lấy chi tiết một gói dịch vụ
      description: |
        Lấy thông tin chi tiết của một gói dịch vụ theo ID.
        **Endpoint public — không cần token.**
      operationId: getServiceById
      security: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Lấy chi tiết thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Admin — Services Management]
      summary: "[Admin] Cập nhật gói dịch vụ"
      description: |
        **[Yêu cầu role: admin]**

        Cập nhật thông tin gói dịch vụ. Chỉ truyền các field muốn thay đổi.
        Có thể dùng để tắt dịch vụ bằng cách đặt `isActive: false`.
      operationId: updateService
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
            example:
              price: 28000
              isActive: true
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Admin — Services Management]
      summary: "[Admin] Xóa gói dịch vụ"
      description: |
        **[Yêu cầu role: admin]**

        Xóa vĩnh viễn gói dịch vụ khỏi hệ thống.

        > ⚠️ **Lưu ý:** Nên cân nhắc dùng `PUT /services/{id}` với `isActive: false`
        > thay vì xóa hẳn, để không ảnh hưởng đến lịch sử đơn hàng đã có.
      operationId: deleteService
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Xóa thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Xóa dịch vụ thành công"
                  data:
                    nullable: true
                    example: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────────────────────────────────────
  #  ORDERS ENDPOINTS  (Base: /api/orders)
  # ─────────────────────────────────────────────────────────────────────────────
...
  /payments/create:
    post:
      tags: [Payments]
      summary: Tạo giao dịch thanh toán
      description: |
        Tạo giao dịch thanh toán cho đơn hàng.

        **Các phương thức hỗ trợ:**
        - `cash`: Thanh toán tiền mặt khi giao đồ. Response không có `paymentUrl`.
        - `vnpay`: Thanh toán online qua VNPay. Response có `paymentUrl` → FE redirect khách tới URL đó.
        - `momo`: Thanh toán qua MoMo. Response có `paymentUrl`.

        **Luồng VNPay:**
        1. FE gọi `POST /payments/create` với `method: "vnpay"`
        2. Server tạo Payment record và trả về `paymentUrl`
        3. FE redirect khách đến `paymentUrl`
        4. VNPay callback về `GET /payments/vnpay-return` (server xử lý tự động)
        5. FE poll `GET /payments/order/{orderId}` để kiểm tra kết quả
      operationId: createPayment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
            example:
              orderId: "64f1a2b3c4d5e6f7a8b9c0d3"
              method: "vnpay"
      responses:
        '201':
          description: Tạo giao dịch thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Tạo giao dịch thanh toán thành công"
                  data:
                    type: object
                    properties:
                      payment:
                        $ref: '#/components/schemas/Payment'
              examples:
                vnpay:
                  summary: Thanh toán VNPay
                  value:
                    success: true
                    message: "Tạo giao dịch thanh toán thành công"
                    data:
                      payment:
                        _id: "64f1a2b3c4d5e6f7a8b9c0d6"
                        order: "64f1a2b3c4d5e6f7a8b9c0d3"
                        amount: 87500
                        method: "vnpay"
                        status: "pending"
                        paymentUrl: "https://sandbox.vnpayment.vn/paymentv2/vpcpay.html?vnp_Amount=8750000&vnp_TxnRef=LD-20240620-0001&..."
                        paidAt: null
                cash:
                  summary: Thanh toán tiền mặt
                  value:
                    success: true
                    message: "Tạo giao dịch thanh toán thành công"
                    data:
                      payment:
                        _id: "64f1a2b3c4d5e6f7a8b9c0d7"
                        order: "64f1a2b3c4d5e6f7a8b9c0d3"
                        amount: 87500
                        method: "cash"
                        status: "pending"
                        paymentUrl: null
                        paidAt: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /payments/order/{orderId}:
    get:
      tags: [Payments]
      summary: Lấy thông tin thanh toán theo đơn hàng
      description: |
        Lấy chi tiết giao dịch thanh toán liên kết với một đơn hàng.
        FE dùng endpoint này để kiểm tra kết quả sau khi khách hoàn tất thanh toán VNPay/MoMo.
      operationId: getPaymentByOrder
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          description: ID của đơn hàng
          schema:
            type: string
            example: "64f1a2b3c4d5e6f7a8b9c0d3"
      responses:
        '200':
          description: Lấy thông tin thanh toán thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Lấy thông tin thanh toán thành công"
                  data:
                    type: object
                    properties:
                      payment:
                        $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/vnpay-return:
    get:
      tags: [Payments]
      summary: VNPay Callback URL (Server tự xử lý — FE không gọi trực tiếp)
      description: |
        Endpoint callback từ VNPay sau khi khách hoàn tất thanh toán.

        > 🤖 **FE không cần gọi endpoint này.** VNPay server tự redirect về đây.
        > Sau khi server xử lý xong, trang sẽ redirect về FE thành công/thất bại.

        **Không yêu cầu JWT** vì VNPay gọi từ server của họ.
      operationId: vnpayReturn
      security: []
      parameters:
        - name: vnp_ResponseCode
          in: query
          schema:
            type: string
          description: Mã kết quả từ VNPay ("00" = thành công)
        - name: vnp_TxnRef
          in: query
          schema:
            type: string
          description: Mã đơn hàng (orderCode)
        - name: vnp_Amount
          in: query
          schema:
            type: string
          description: Số tiền thanh toán (×100)
      responses:
        '302':
          description: Redirect về trang FE (thành công hoặc thất bại)

  # ─────────────────────────────────────────────────────────────────────────────
  #  ADMIN ENDPOINTS  (Base: /api/admin)
  # ─────────────────────────────────────────────────────────────────────────────
...